predict_model/pyproject.toml
[tool.poetry]
name = "model"
version = "0.1.0"
description = ""
authors = ["bluewarn <n5lp97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.3"
pandas = "^2.2.3"
seaborn = "^0.13.2"
matplotlib = "^3.9.2"
scikit-learn = "^1.5.2"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
